requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pinecone-client==5.0.1
openai==1.51.0
PyPDF2==3.0.1
python-docx==1.1.2
#sentence-transformers==3.1.1
python-dotenv==1.0.1
aiohttp==3.10.5
httpx==0.27.0
gunicorn==22.0.0
psycopg2-binary>=2.9.9